+++ packages/gamut/src/GridForm/GridFormSubmit/index.tsx
@@ -3,23 +3,40 @@ import styled from '@emotion/styled';
 import React from 'react';
 
 import { Box } from '../../Box';
+import { CTAButton, FillButton } from '../../Button';
 import {
   ButtonDeprecated,
   ButtonDeprecatedProps,
 } from '../../ButtonDeprecated';
 import { Column, ColumnSizes } from '../../Layout';
+import { VisualTheme } from '../../theming/VisualTheme';
 import { ResponsiveProperty } from '../../typings/responsive-properties';
 
 export type GridFormSubmitPosition = keyof typeof positions;
 
-export type GridFormSubmitProps = {
+export type ButtonType = 'cta' | 'fill';
+export type ButtonDeprecatedType = 'business';
+
+type GridFormBase = {
   contents: React.ReactNode;
-  disabled?: ButtonDeprecatedProps['disabled'];
   position?: GridFormSubmitPosition;
-  outline?: ButtonDeprecatedProps['outline'];
   size: ResponsiveProperty<ColumnSizes>;
-  theme?: ButtonDeprecatedProps['theme'];
+  disabled?: ButtonDeprecatedProps['disabled'];
+  mode?: VisualTheme;
 };
+
+type GridFormButtonSubmitPropsDeprecated = GridFormBase & {
+  type?: ButtonDeprecatedType;
+  theme?: ButtonDeprecatedProps['theme'];
+  outline?: ButtonDeprecatedProps['outline'];
+};
+
+type GridFormSubmitPropsStandard = GridFormBase & {
+  type?: ButtonType;
+};
+export type GridFormSubmitProps =
+  | GridFormButtonSubmitPropsDeprecated
+  | GridFormSubmitPropsStandard;
 
 const positions = {
   left: 'flex-start',
@@ -30,30 +47,50 @@ const positions = {
 
 const StyledColumn = styled(Column)(flex);
 
-export const GridFormSubmit: React.FC<GridFormSubmitProps> = ({
-  contents,
-  disabled,
-  outline,
-  position = 'left',
-  size,
-  theme = 'brand-purple',
-}) => {
+export const GridFormSubmit: React.FC<GridFormSubmitProps> = (props) => {
+  const getButton = () => {
+    switch (props.type) {
+      case 'cta':
+        return (
+          <CTAButton type="submit" mode={props.mode} disabled={props.disabled}>
+            {props.contents}
+          </CTAButton>
+        );
+      case 'business':
+        /**
+         * There are current designs that currently rely on the deprecated button.
+         * Primarily business components such as /WorkerSupportApplication/index.tsx,
+         * /PlanInvitationBulkForm/PlanInvitationBulkForm.tsx and /PlanInvitationForm/PlanInvitationForm.tsx
+         * currently using brand-blue variant of the deprecated button. With the later two also using the also
+         * using the outline button. Once work is finished for the colorMode changes for buttons the deprecated button
+         * can be removed.
+         */
+        return (
+          <ButtonDeprecated
+            disabled={props.disabled}
+            outline={props.outline}
+            theme={props.theme}
+            type="submit"
+          >
+            {props.contents}
+          </ButtonDeprecated>
+        );
+      default:
+        return (
+          <FillButton type="submit" mode={props.mode} disabled={props.disabled}>
+            {props.contents}
+          </FillButton>
+        );
+    }
+  };
+
   return (
     <StyledColumn
-      justifyContent={positions[position]}
+      justifyContent={positions[props.position || 'left']}
       alignItems="center"
-      size={size}
+      size={props.size}
     >
-      <Box marginBottom={8}>
-        <ButtonDeprecated
-          disabled={disabled}
-          outline={outline}
-          theme={theme}
-          type="submit"
-        >
-          {contents}
-        </ButtonDeprecated>
-      </Box>
+      <Box marginBottom={8}>{getButton()}</Box>
     </StyledColumn>
   );
 };