+++ packages/gamut/CHANGELOG.md
@@ -3,6 +3,75 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [29.7.1](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut@29.7.0...@codecademy/gamut@29.7.1) (2021-04-05)
+
+**Note:** Version bump only for package @codecademy/gamut
+
+
+
+
+
+## [29.7.0](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut@29.6.0...@codecademy/gamut@29.7.0) (2021-04-01)
+
+
+### Features
+
+* **Input:** React-Recurly edits ([1d2a286](https://github.com/Codecademy/client-modules/commit/1d2a28670b6b218a891462191a80e97a982482c9))
+
+
+
+## [29.6.0](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut@29.5.2...@codecademy/gamut@29.6.0) (2021-04-01)
+
+
+### Features
+
+* **GridForm:** Update the buttons for GridForm submit ([cb03dd2](https://github.com/Codecademy/client-modules/commit/cb03dd2a7bce0d2b9baa99349bae22824c0e614b))
+
+
+
+### [29.5.2](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut@29.5.1...@codecademy/gamut@29.5.2) (2021-03-30)
+
+
+### Bug Fixes
+
+* **NotificationListNew:** remove top pattern ([84c7750](https://github.com/Codecademy/client-modules/commit/84c775054992491e57a8b82f62365517fd216c78))
+
+
+
+### [29.5.1](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut@29.5.0...@codecademy/gamut@29.5.1) (2021-03-30)
+
+**Note:** Version bump only for package @codecademy/gamut
+
+
+
+
+
+## [29.5.0](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut@29.4.8...@codecademy/gamut@29.5.0) (2021-03-30)
+
+
+### Features
+
+* **Text:** Deprecate Text / Heading and add more robust UI Text Component [GM-162] ([78e284f](https://github.com/Codecademy/client-modules/commit/78e284f3ce80f14e583af3133f692a7cd91e2196))
+
+
+
+### [29.4.8](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut@29.4.7...@codecademy/gamut@29.4.8) (2021-03-30)
+
+**Note:** Version bump only for package @codecademy/gamut
+
+
+
+
+
+### [29.4.7](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut@29.4.6...@codecademy/gamut@29.4.7) (2021-03-29)
+
+
+### Bug Fixes
+
+* pattern should have aria-hidden ([10b87be](https://github.com/Codecademy/client-modules/commit/10b87bee2b9f3a9a261f3560f28a7295b91bb7d4))
+
+
+
 ### [29.4.6](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut@29.4.5...@codecademy/gamut@29.4.6) (2021-03-26)
 
 **Note:** Version bump only for package @codecademy/gamut