+++ packages/konamimojisplosion/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [0.2.4](https://github.com/Codecademy/client-modules/compare/@codecademy/konamimojisplosion@0.2.3...@codecademy/konamimojisplosion@0.2.4) (2021-04-05)
+
+**Note:** Version bump only for package @codecademy/konamimojisplosion
+
+
+
+
+
+### [0.2.3](https://github.com/Codecademy/client-modules/compare/@codecademy/konamimojisplosion@0.2.2...@codecademy/konamimojisplosion@0.2.3) (2021-03-30)
+
+**Note:** Version bump only for package @codecademy/konamimojisplosion
+
+
+
+
+
+### [0.2.2](https://github.com/Codecademy/client-modules/compare/@codecademy/konamimojisplosion@0.2.1...@codecademy/konamimojisplosion@0.2.2) (2021-03-30)
+
+**Note:** Version bump only for package @codecademy/konamimojisplosion
+
+
+
+
+
 ### [0.2.1](https://github.com/Codecademy/client-modules/compare/@codecademy/konamimojisplosion@0.2.0...@codecademy/konamimojisplosion@0.2.1) (2021-03-26)
 
 **Note:** Version bump only for package @codecademy/konamimojisplosion